+++ src/includes/classes/App.php
@@ -177,7 +177,7 @@ class App extends Classes\Core\Base\Core
      *
      * @type string Version.
      */
-    const VERSION = '160719.52827'; //v//
+    const VERSION = '160720.50539'; //v//
 
     /**
      * Constructor.