+++ src/includes/classes/Core/Utils/Html2Text.php
@@ -62,11 +62,11 @@ class Html2Text extends Classes\Core\Base\Core implements Interfaces\HtmlConstan
         $string = $this->c::unescHtml($string);
         $string = str_replace("\xC2\xA0", ' ', $string);
 
-        if ($br2nl) {
+        if ($br2nl) { // Default behavior.
             $string = preg_replace('/\<br(?:\>|\/\s*\>|\s[^\/>]*\/\s*\>)/u', "\n", $string);
             $string = $this->c::normalizeEols($string); // Normalize line breaks.
             $string = preg_replace('/[ '."\t\x0B".']+/u', ' ', $string);
-        } else {
+        } else { // Force everything into a single line.
             $string = preg_replace('/\s+/u', ' ', $string);
         }
         $string = $this->c::mbTrim($string);